refactor: move logic
